renaming of tweak files
